+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Add `/accounts` api endpoint
 
+### Changed
+ - Api now accepts `S: Into<String>`
+
 
 ## [0.4.3] - 2020-07-24
 ### Changed